requirements.txt
pytz
logging_tree
html5lib==0.999999999  # last version with "html5lib.sanitizer" required (or not anymore?) by djangocms_text_ckeditor
Pillow>=2.7.0
six
decorator==4.4.0
PyYAML==5.1
beautifulsoup4==4.7.1
python-memcached==1.59

Django==2.1.11  # can't switch to 2.2.1 until PyMysql gets updated to match MySQLdb

django-compat-patcher==0.5

django-environ==0.4.5
django-cms==3.6.0
django-treebeard==4.3.1
easy_thumbnails==2.7
django-polymorphic==2.0.3
django-filer==1.5.0
django-compressor==2.4
django-request==1.5.5
django-extensions==2.2.8
django-taggit==1.2.0

djangocms-column==1.9.0
djangocms-googlemap==1.4.0
git+https://github.com/isotoma/djangocms-inherit@b19735ea2a5b4552eb6e619422b0a7d8dd67da06
djangocms-link==2.4.0
djangocms-style==2.3.0
djangocms-snippet==2.3.0
cmsplugin-filer==1.1.3
djangocms-admin-style==1.4.0
djangocms-text-ckeditor==3.7.0
git+https://github.com/pakal/cmsplugin-rst/
https://github.com/makukha/cmsplugin-raw-html/archive/master.zip
djangocms-page-tags==0.8.0
djangocms-page-sitemap==0.5.3

django-terms==1.2.0
django-newsletter==0.9
django-tinymce==2.8.0

django-reversion==3.0.3
django-mptt==0.10.0
django-classy-tags==1.0.0
django-sekizai==1.1.0

# Drop-in replacement for MySQLdb, when using its install_as_MySQLdb() patcher
PyMySQL==0.9.3

# for development
django-debug-toolbar==1.11
